Reorganize item packages.
