+++ src/main/java/info/coremodding/craftenchants/item/ItemCE.java
@@ -3,7 +3,7 @@
  * http://www.gnu.org/licenses/gpl.html
  * http://www.gnu.org/licenses/lgpl.html
  */
-package info.coremodding.craftenchants.enchants;
+package info.coremodding.craftenchants.item;
 
 import info.coremodding.craftenchants.library.Reference;
 import info.coremodding.craftenchants.proxy.Proxy;
