+++ src/main/java/info/coremodding/craftenchants/item/enchants/Counterweight.java
@@ -3,9 +3,9 @@
   * http://www.gnu.org/licenses/gpl.html
   * http://www.gnu.org/licenses/lgpl.html
   */
-package info.coremodding.craftenchants.enchants.item.enchants;
+package info.coremodding.craftenchants.item.enchants;
 
-import info.coremodding.craftenchants.enchants.ItemCE;
+import info.coremodding.craftenchants.item.ItemCE;
 
 import net.minecraft.enchantment.Enchantment;
 import net.minecraft.item.EnumToolMaterial;
