+++ src/main/java/info/coremodding/craftenchants/item/recipes/VanillaRecipes.java
@@ -2,12 +2,12 @@
  * Copyright 2014 by CoreModding under GNU Lesser General Public License (LGPLv3)
  * http://www.gnu.org/licenses/gpl.html http://www.gnu.org/licenses/lgpl.html
  */
-package info.coremodding.craftenchants.enchants.item.recipes;
+package info.coremodding.craftenchants.item.recipes;
 
 import net.minecraft.item.ItemStack;
 import cpw.mods.fml.common.registry.GameRegistry;
 
-import info.coremodding.craftenchants.enchants.ItemsCE;
+import info.coremodding.craftenchants.item.ItemsCE;
 import info.coremodding.craftenchants.library.ModRecipes;
 
 public class VanillaRecipes {
